feat: make maintain loop more reusable
